Merges latest master prior to finalizing release
In case a preivous build modified master while this build was running,
we want to merge those changes before running this finalize.  Even if
the changes on master were significant, since finalizing builds are
based on blobs stored in s3 there is no harm in having these changes.
The worst case is that the git logs may not represent the changes in
each release.

[#95585962](https://www.pivotaltracker.com/story/show/95585962)
Signed-off-by: Aaron Hurley <ahurley@pivotal.io>