Changes to a manually triggered pipeline
...with no develop->master promotion. This will allow a simpler build
pipeline with less race conditions to try to work through.

Also rebases the branch when pushing it back up in case more changes
were made while this build ran, we want to ensure we can push up the
final release state.

Note: it will be up to the operator to not kick off a new build when one
is running.

[#95585962](https://www.pivotaltracker.com/story/show/95585962)
Signed-off-by: Aaron Hurley <ahurley@pivotal.io>