Bumps bosh_openstack_cpi gem to unpublished version
We want to test some src changes to the gem before we publish the gem.
The best way to test them is through this releases ci pipeline so we are
committing a temporary unpublished gem.  We should not cut a release
using this gem, but instead wait for a published gem be made and
re-added to this project

[#100433378](https://www.pivotaltracker.com/story/show/100433378)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>