Added a cleanup job that runs exclusively of the main pipeline
As part of this commit, a pool resource is used to guarantee that
the cleanup job does not run while the main pipeline is running
or vice versa. A new job is added to the end of the pipeline
that releases this lock when the main pipeline is finished up
to the promote-candidate job. We also created separate Bluebox
users for bats and lifecycles instead of using the admin user which
has permissions to all projects under that account.

[#104770254](https://www.pivotaltracker.com/story/show/104770254)

Signed-off-by: Edwin Xie <exie@pivotal.io>