Get rid of global variable in prepare_bosh_release
Building upon the previous commit this adds distro as an arg
to the test-upgrade task and the corresponding prepare_bosh_release
invocation. The distro value was already set in the main pipeline
manifest of certify-stemcell.

Now every invocation of prepare_bosh_release uses arguments and
the global variable distro, which was hiding away in line 46,
can be removed.