Fix ordering of OpenStack calls executed by `delete_vm`
LBaaS pool membership must be deleted before VM is deleted.
Otherwise, the loadbalancer will keep forwarding requests to the
not existing VM until its health monitor recognises the VM as `down`.

LBaaS pool membership needs to be deleted before port is deleted.
Otherwise, OpenStack will recreate a port for the IP which is associated
to the membership.

[#148837971](https://www.pivotaltracker.com/story/show/148837971)

Signed-off-by: Mario Manno <mario.manno@gmail.com>