Fix verification failures of cpi to openstack.
The bug results in failing verifications of the cpi to openstack because
connection_options are missing in the authentication process. The
connection_options are now directly passed to Excon.defaults.

Has to be removed as soon as a new fog-openstack version is released,
containing a fix.

Issue: https://github.com/fog/fog-openstack/issues/441
PR fixing the Issue: https://github.com/fog/fog-openstack/pull/442

Co-authored-by: Denis Langer <denis.langer@sap.com>