Fix shebang so that it points to a valid ruby binary
As of rubygems 2.7.6 and ruby 2.4.4 they fixed a vulnerability with regards
to path traversal and symlinks. The resulting fix uses the real path for the
shebang which includes a checksum that is not constant across BOSH vms.

We needed to modify the shebang of the gem and bundle scripts to use a path containing ${BOSH_INSTALL_TARGET}.

[finishes #160883099](https://www.pivotaltracker.com/story/show/160883099)

Co-authored-by: Max Becker <max.becker@sap.com>