Assure that non-default volume type is used for volume_spec
Fixes a bug that occured when only the default volume type was available in openstack.
It led to tests in volume spec not checking that volume type in
cloud_properties section (or openstack.default_volume_type in the global
CPI config) is used.

Also changes the use of the attribute volume_type from the lifecycle.yml:
Removed the prior usage because it was unnecessary used in lifecycle_spec.
Now it is a mandatory attribute for the volume_spec.

This requires from now on all Openstacks to be setup with a custom volume_type for
the lifecycle test.

Co-authored-by: David Ansari <david.ansari@sap.com>