Filter for OpenStack project
If the user running the CFOV has an admin role on the project
the validator runs in, the cleanup script will find all glance
images and all neutron ports.
The list commands filters by default on the project for servers, volumes
and snapshot.
This patch adds the project filter explicitely for images and ports.
If we cannot retrieve the OpenStack project ID, we exit. Because we
would delete to many ports, if run as admin user.

Additionally we filter the neutron port list for DVR ports.
(Maybe a whitelist for device_owner=compute:* is better here?)