Add registry-less test cases to cpi pipeline
Adds additional jobs to the cpi pipeline to test compatibility with
stemcell/director combinations with and without registry.
Also adds ops files to move agent properties to `env/bosh`.

Remark: To test registry less deployment we needed a stemcell that supports API_Version 2. Therefore, a custom stemcell is used which is build from the agent branch `registry-removal`. Once the agent changes are merged into master the pipeline needs to be adapted.

[#161548774](https://www.pivotaltracker.com/story/show/161548774)

Co-authored-by: Vladimir Videlov <vladimir.videlov@sap.com>
Co-authored-by: Max Becker <max.becker@sap.com>
Co-authored-by: Denis Holdeew <denis.holdeew@sap.com>
Co-authored-by: Beyhan Veli <beyhan.veli@sap.com>