Use terraform string templates for allocation pools
Before this change the terraform output of `primary_net_dhcp_pool`
and `secondary_net_dhcp_pool` was null, due to a misconfiguration.
The `openstack_networking_subnet_v2.allocation_pool` property is
actually a set rather than a list which caused the problem when
accessing the start and end properties of the allocation pool.