Exclude SSH tests as internal env is limited
In the openstack env used against the current pipelines, it's tucked
away very securely. An instance that needs to be bosh ssh'd (or just
ssh'd) into needs to be deployed with a floating IP. Unfortunately, when
bosh ssh is invoked, the agent gives the internal/non-floating IP as the
host. The cli then attempts to ssh with the internal address, which
never succeeds. One idea could be to change the CLI to allow the
operator to override the host. This likely isn't worth it, as this
situation doesn't happen much in real foundations.