Adds Task to print config file repo
This commit adds a Concourse task that -when run- *should* print out the
repo that contains the pipeline's config file, as well as the filename
of the pipeline config file.

Unfortunately, both the repo containing the file as well as the file's
name are not automatically determined, but this _is_ substantially better
than what we had before (which was nothing).

[#179855722] As a Bosh Ecosystem developer, I would like to be able to discover which repo backs a Concourse Pipeline