Adds Openstack Wallaby CA cert
This will be needed by Terraform to talk with the VMware Openstack
Wallaby cluster. Ordinarily, I'd be hesitant to add company-specific
resources to an open source repo, but we already have the
`ci/configure.sh` script which is tailored for our internal environment.
So there's precedent for doing just that...

[#179890893] Migrate the OpenStack CPI CI to the new Openstack (once OpenStack works).