Pass Openstack CA cert to Director
The Bosh Director also needs to have access to the CA certificate of the
Openstack cluster. This commit should set things up so that the Director
now has that cert.

This should be designed so that you can pass the empty string in the
`openstack_ca_cert` environment variable and it will behave like it used
to, but I haven't actually integration tested it yet. We'll see!

[#179890893] Migrate the OpenStack CPI CI to the new Openstack (once OpenStack works).