🐞 Use `mkpasswd`, not Ruby, to create password hash
Using Ruby's `String#crypt` to generate a SHA-512 password doesn't work
on Focal Fossil (it generates "*0" instead of a real password hash), and
the Ruby developers have deprecated that method for many good reasons.

We now use `mkpasswd` as an alternative.

Note that we use exactly 10 bytes as a salt to the SHA-512 hashing
algorithm. This avoids `base32` padding the end of the salt with an "=",
which is illegal. `base32` converts 10 bytes to exactly 16 characters,
no padding.

Fixes:
```
invalid argument for flag `-v, --var' (expected []template.VarKV): Deserializing variables 'bosh_vcap_password_hash=*0': yaml: unknown anchor '0' referenced
```