Adds empty bundler config for CI
Because of the changes in 9eda590 CI is now
complaining about this when running the CPI:

```
D, [2022-02-24T19:21:33.889524 #32707] [task:1] DEBUG -- DirectorJobRunner: [external-cpi] [cpi-244903] response: , err: There was an error while trying to write to
`/var/vcap/packages/bosh_openstack_cpi/.bundle/config`. It is likely that you
need to grant write permissions for that path.
, exit_status: 23
```

_Hopefully_, creating this config file will give `bundle` a file that
has write permissions for whoever it's running as in CI so that it can
edit it as it now needs to.

Signed-off-by: Kenneth Lakin <klakin@vmware.com>