Fix Gemfile.lock
It appears that the prior commit which updated `fog-openstack` to 1.1.0
did not do so by executing `bundle update fog-openstack`. When one
explicitly runs this command, the Gemfile.lock pulls many updates to
match the fog-openstack dependencies (including the removal of
ipaddress, which has caused some odd problems in environments without
internet connectivity). This commit checks in the result of Gemfile.lock
after running the proper `bundle update` command.