Updates Openstack Wallaby CA cert
With the move to the LVN datacenter, many things had to be changed in
the Openstack Wallaby cluster. The CA cert is one of those things.

This commit SHOULD make the Openstack CPI pipelines run correctly
against the LVN Openstack Wallaby cluster... once the IT firewall
modification ticket gets done.