Uses Docker Image resource for image source
This commit switches from using the docker image specified in the Task
.YML file (which queries DockerHub every time the task is run) to the
docker image specified in the `resources` array in the pipeline
definition for the 'write_ca_file' and 'test' Tasks.

Why? We've observed that when the 'lifecycle' Job is run multiple times
in one hour, the 'write_ca_file' and 'test' Tasks will fail early
because DockerHub refuses the pre-task-start image check request due to
rate limiting.  Because of how Concourse is designed, this check happens
every time the Task is run. By switching to a preexisting resource
that's configured to check on a specified schedule, we hope to avoid
Task failures due to DockerHub rate limiting.

Why are we not making this change to the other two Tasks in the Job?
Because a later commit will make major changes to those Tasks.