Switches away from Terraform Resource to script
This commit switches the 'lifecycle' Job away from the Terraform
Resource to a simple pair of scripts that run 'terraform apply' and
'terraform destroy'. (It also adds the 'openstack' tag to image
resources that are being used by Tasks.)

Switching away from the Terraform Resource means that the .tfstate file
must be passed around between Tasks, but given that we only have one
'openstack' tagged worker, this is not a large burden. It also means
that we lose much of the configuration flexibility that the Terraform
Resource offers us, but we're making use of none of that in the
'lifecycle' Job.

Why are we making this change?

When running the 'lifecycle' Job in CI, we often see mysterious
failures. We've seen failures after 'terraform destroy' where the
Resource complains that the Terraform Workspace is non-empty and fails
the operation... except that the 'terraform destroy' successfully
destroyed everything that Terraform created. We've seen failures before
'terraform apply' where the Terraform Resource detonates because the
.tfstate file it's looking for is not present in the GCS bucket it has
been instructed to use. (This is despite the fact that this should not
be a fatal error for the Resource). We've also seen failures where the
contents of the "IDs of the resources Terraform has provisioned"
metadata file passed along to the 'test' task contains nearly none (or
sometimes absolutely none) of the IDs of the resources that were
actually provisioned.... causing both the tests and the subsequent
'terraform destroy' to fail to function correctly. Testing has revealed
that converting to a very simple pair of scripts bypasses all of these
problems.

There are some open questions:
1) Why are these errors intermittent?
2) Why do they often happen to the 'lifecycle' Job and nearly never to
   the 'bats-ubuntu-manual' Job?

We've burned a LOT of time trying to answer these questions and have
given up on getting answers. Definitely feel free to start your own
investigation and switch back to the Terraform Resource if you figure
out what's going wrong and how to avoid it.