Moves bats-ubuntu-manual Job to Terraform script
This commit moves the bats-ubuntu-manual Job to the same sort of
terraform apply/destroy script that's working very well for the
lifecycle Job.

Why are we doing this? Because the last six runs of the lifecycle Job
have all succeeded, whereas the last six runs of the bats-ubuntu-manual
Job have failed due to Terraform "workspace" errors about half of the
time. Moving to this script seems to get rid of any obvious usage of the
Terraform "workspace" stuff, which seems to give our Builds a much, much
higher success rate. (The fact that we don't need to transfer anything
to or from GCS with this way of doing things is just a nice bonus.)