Handle 'tolist' Terraform construction
This commit should correctly handle the 'tolist' construction in the
output from the 'terraform output' command. That construction is a bit
too complex for my 'sed' skills, so I've fallen back on my crummy 'awk'
skills to provide a replacement to the inline 'sed'.

The 'awk' program definitely could be improved, but it's fit for its
current limited purpose.